Consul/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.4
django-cors==0.1
django-cors-headers==4.6.0
django-filter==24.3
django-modelcluster==6.4
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.1
msal==1.31.1
oauthlib==3.2.2
openpyxl==3.1.5
pillow==11.1.0
pillow_heif==0.21.0
pycparser==2.22
PyJWT==2.10.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
urllib3==2.3.0
wagtail==6.3.2
wagtail-headless-preview==0.8.0
Willow==1.9.0
